Editted some classes

+++ src/main/java/com/example/ProjectManagementSystem/services/impl/UserServiceImpl.java
@@ -1,12 +1,14 @@
 package com.example.ProjectManagementSystem.services.impl;
 
 
+import com.example.ProjectManagementSystem.dtos.Requests.LoginRequest;
 import com.example.ProjectManagementSystem.dtos.Requests.RegisterRequest;
 import com.example.ProjectManagementSystem.exceptions.UserNotFoundException;
 import com.example.ProjectManagementSystem.models.User;
 import com.example.ProjectManagementSystem.repositories.UserRepository;
 import com.example.ProjectManagementSystem.services.UserService;
 import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -15,15 +17,30 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.*;
 
 @Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class UserServiceImpl implements UserService {
-    @Autowired private UserRepository repo;
-    @Autowired private PasswordEncoder encoder;
+    @Autowired private final UserRepository repo;
+    @Autowired private final PasswordEncoder encoder;
 
+
+@Override
+public User register(@Valid RegisterRequest request) {
+    User user = new User();
+    user.setName(request.getName());
+    user.setEmail(request.getEmail());
+    user.setPasswordHash(encoder.encode(request.getPasswordHash()));
+    return repo.save(user);
+}
 
     @Override
-    public User register(@Valid RegisterRequest user) {
-        user.setPasswordHash(encoder.encode(user.getPasswordHash()));
-        return repo.save(user);
+    public User login(@Valid LoginRequest request) {
+        Optional<User> userOpt = repo.findByEmail(request.getEmail());
+        if (userOpt.isPresent() && encoder.matches(request.getPassword(), userOpt.get().getPasswordHash())) {
+            return userOpt.get();
+        } else {
+            throw new UserNotFoundException("Invalid email or password");
+        }
     }
 
     @Override
@@ -44,6 +61,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteUser(Long userId) {
+        if(!repo.existsById(userId)){
+            throw new UserNotFoundException("Cannot delete, user not found with ID " + userId);
+        }
         repo.deleteById(userId);
     }
